*fixing reorganizing config package, documentations and adjusting errors import
