+++ src/domain/Convenio.ts
@@ -1,6 +1,6 @@
 import { DataTypes, InferAttributes, InferCreationAttributes, Model } from "sequelize";
-import sequelize from "../../database/postgresqlConfig";
-import Ifes from "../models/Ifes";
+import sequelize from "../config/postgresqlConfig";
+import Ifes from "../domain/Ifes";
 import Convenente from "./Convenente";
 
 class Convenio extends Model<InferAttributes<Convenio>, InferCreationAttributes<Convenio>> {
