+++ src/domain/Ifes.ts
@@ -1,5 +1,5 @@
 import { DataTypes, InferAttributes, InferCreationAttributes, Model } from "sequelize";
-import sequelize from "../../database/postgresqlConfig";
+import sequelize from "../config/postgresqlConfig";
 import Convenio from "../domain/Convenio";
 
 class Ifes extends Model<InferAttributes<Ifes>, InferCreationAttributes<Ifes>> {
